add summary func

+++ llm-project-nlp/task_ner/run_bert_bilstm_crf.py
@@ -5,6 +5,7 @@
 # @function:
 import logging
 import hydra
+import torch
 from omegaconf import DictConfig, OmegaConf
 from torch.utils.data import DataLoader
 from torchmetrics import F1Score, MetricCollection, Accuracy
@@ -60,9 +61,8 @@ def train(config: DictConfig):
     logging.info("初始化模型")
     model: BertBiLstmCrf = BertBiLstmCrf(pretrain_path=config.pretrain_path, num_labels=num_labels, dropout=config.dropout,
                                          lstm_num_layers=config.lstm_num_layers, lstm_hidden_size=config.lstm_hidden_size)
-    print(summary(model, input_shape=[(config.max_seq_len, ),
-                                      (config.max_seq_len, ),
-                                      (config.max_seq_len, )]))
+    print(summary(model, input_shape=[(config.max_seq_len, ), (config.max_seq_len, ), (config.max_seq_len, )],
+                  input_dtype=torch.LongTensor))
 
     logging.info("配置优化器、学习率调整器、损失函数、评估指标")
     optimizer = build_optimizer(model, learning_rate=config.learning_rate)
